Introduce DependencyTracker::IncrementalDependency
Promote the simple path list to a list of path-fingerprint pairs. The idea being that we'll write these out in the swiftdeps files serialized in the swiftmodule file. When the driver deserializes the node, it can now read the fingerprint to determine whether the top-level contents of the module have changed. This is essentially providing an interface hash for serialized Swift modules.