Add Module-Level Interface Hash Serialization
The interface hash of a module is akin to the interface hash for an SourceFile. That is, a change in the interface hash signals that a change to a top-level declaration *somewhere* in the module has been made. Quite where that change occurred is still something the driver has to figure out for itself.

However, we have to be sensitive to the many representations a Swift module can have, and how those representations are generated. In the easy case, we're in incremental (batch) mode. We write down a list of interface hashes from each of our primaries into the module and keep going. However, in order to complete merge-modules, we need the notion of a "unified" interface hash for each of the files. This is the tricky case: when any of the files in the module is serialized.

To define a stable interface hash for a module composed of many serialized sub-units:

* For each sub-unit, extract its list of interface hashes.
* Sort these interface hashes - lexicographically will do - to remove non-determinism inherent in batch mode
* Combine each of these hashes

The resulting hash is invariant with respect to the ordering of primary hashes. And to satisfy the avalanche invariant of Fingerprints, will change whenever any of the hashes of any of the primaries changes at any time.