Teach the Driver to Check The Interface Hash of Incremental External Dependencies
Add a quite conservative check for the interface hash of incremental external dependencies. We require that each dependent has written down an interface hash that exactly matches the current interface hash of the deserialized module's  interface node in the dependency graph. When there is a mismatch between the fingerprint the primary file has written down in its dependency graph and the fingerprint for the deserialized swiftmodule we now have a clear signal that we need to recompile the dependent.

This is equivalent to the interface hash changing for a normal Swift file.

rdar://72113278